+++ .env.example
@@ -4,7 +4,7 @@ DATABASE_URL="postgresql://user:password@localhost:5432/tanstarter"
 # You can also use Docker Compose to set up a local PostgreSQL database:
 # docker-compose up -d
 
-# https://www.better-auth.com/docs/installation
+# pnpm run auth:secret
 BETTER_AUTH_SECRET=
 
 # OAuth2 Providers, optional
