+++ package.json
@@ -11,8 +11,8 @@
     "db": "drizzle-kit",
     "deps": "pnpm dlx npm-check-updates@latest --interactive --format group",
     "ui": "pnpm dlx shadcn@latest",
-    "auth": "pnpm dlx @better-auth/cli@latest",
-    "auth:generate": "pnpm dlx @better-auth/cli@latest generate --y --output ./src/lib/server/schema/auth.schema.ts && prettier --write ./src/lib/server/schema/auth.schema.ts"
+    "auth:secret": "pnpm dlx @better-auth/cli@latest secret",
+    "auth:generate": "pnpm dlx @better-auth/cli@latest generate --config ./src/lib/auth/index.ts --y --output ./src/lib/db/schema/auth.schema.ts && prettier --write ./src/lib/db/schema/auth.schema.ts"
   },
   "dependencies": {
     "@radix-ui/react-label": "^2.1.6",
@@ -42,7 +42,7 @@
     "@tanstack/eslint-plugin-query": "^5.74.7",
     "@tanstack/eslint-plugin-router": "alpha",
     "@types/node": "^22.15.17",
-    "@types/react": "^19.1.3",
+    "@types/react": "^19.1.4",
     "@types/react-dom": "^19.1.4",
     "babel-plugin-react-compiler": "latest",
     "drizzle-kit": "^0.31.1",
@@ -55,7 +55,7 @@
     "tailwindcss": "^4.1.6",
     "tw-animate-css": "^1.2.9",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.32.0",
+    "typescript-eslint": "^8.32.1",
     "vite-tsconfig-paths": "^5.1.4"
   }
 }
