+++ src/components/ThemeToggle.tsx
@@ -1,5 +1,5 @@
 import { MoonIcon, SunIcon } from "lucide-react";
-import { Button } from "./ui/button";
+import { Button } from "~/components/ui/button";
 
 export default function ThemeToggle() {
   function toggleTheme() {
