+++ src/lib/auth/index.ts
@@ -1,0 +1,45 @@
+import { serverOnly } from "@tanstack/react-start";
+import { betterAuth } from "better-auth";
+import { drizzleAdapter } from "better-auth/adapters/drizzle";
+import { reactStartCookies } from "better-auth/react-start";
+
+import { db } from "~/lib/db";
+
+const getAuthConfig = serverOnly(() =>
+  betterAuth({
+    baseURL: process.env.VITE_BASE_URL,
+    database: drizzleAdapter(db, {
+      provider: "pg",
+    }),
+
+    // https://www.better-auth.com/docs/integrations/tanstack#usage-tips
+    plugins: [reactStartCookies()],
+
+    // https://www.better-auth.com/docs/concepts/session-management#session-caching
+    // session: {
+    //   cookieCache: {
+    //     enabled: true,
+    //     maxAge: 5 * 60, // 5 minutes
+    //   },
+    // },
+
+    // https://www.better-auth.com/docs/concepts/oauth
+    socialProviders: {
+      github: {
+        clientId: process.env.GITHUB_CLIENT_ID!,
+        clientSecret: process.env.GITHUB_CLIENT_SECRET!,
+      },
+      google: {
+        clientId: process.env.GOOGLE_CLIENT_ID!,
+        clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+      },
+    },
+
+    // https://www.better-auth.com/docs/authentication/email-password
+    emailAndPassword: {
+      enabled: true,
+    },
+  }),
+);
+
+export const auth = getAuthConfig();
