+++ src/lib/auth/middleware/auth-guard.ts
@@ -1,6 +1,6 @@
 import { createMiddleware } from "@tanstack/react-start";
 import { getWebRequest, setResponseStatus } from "@tanstack/react-start/server";
-import { auth } from "~/lib/server/auth";
+import { auth } from "~/lib/auth";
 
 // https://tanstack.com/start/latest/docs/framework/react/middleware
 // This is a sample middleware that you can use in your server functions.
