+++ src/lib/db/index.ts
@@ -1,0 +1,13 @@
+import { serverOnly } from "@tanstack/react-start";
+import { drizzle } from "drizzle-orm/postgres-js";
+import postgres from "postgres";
+
+import * as schema from "~/lib/db/schema";
+
+const driver = postgres(process.env.DATABASE_URL as string);
+
+const getDatabase = serverOnly(() =>
+  drizzle({ client: driver, schema, casing: "snake_case" }),
+);
+
+export const db = getDatabase();
