+++ src/lib/db/schema/auth.schema.ts
