--- src/lib/server/db.ts
@@ -1,7 +1,0 @@
-import { drizzle } from "drizzle-orm/postgres-js";
-import postgres from "postgres";
-
-import * as schema from "./schema";
-
-const driver = postgres(process.env.DATABASE_URL as string);
-export const db = drizzle({ client: driver, schema, casing: "snake_case" });
