+++ src/routes/(auth)/login.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { Button } from "~/components/ui/button";
 import { Input } from "~/components/ui/input";
 import { Label } from "~/components/ui/label";
-import authClient from "~/lib/auth-client";
+import authClient from "~/lib/auth/auth-client";
 
 export const Route = createFileRoute({
   component: LoginForm,
