+++ src/styles.css
@@ -1,4 +1,4 @@
-@import "tailwindcss" source("../../");
+@import "tailwindcss" source("./");
 
 @import "tw-animate-css";
